requirements.txt
opencv-python
streamlit
Pillow
numpy
matplotlib
sounddevice
mediapipe
torch
torchvision
scipy
streamlit-option-menu
